requirements.txt
python-snippets
agit
pydantic